requirements.txt
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.12.3
Click==7.0
cloudinary==1.17.0
dnspython==1.16.0
Flask==1.1.1
Flask-Assets==0.12
Flask-CKEditor==0.4.3
Flask-Cors==3.0.8
Flask-Dropzone==1.5.4
Flask-FontAwesome==0.1.2
Flask-JWT-Extended==3.23.0
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
mailer==0.8.1
MarkupSafe==1.1.1
mock==3.0.5
mongoengine==0.18.2
pycparser==2.19
PyJWT==1.7.1
pymongo==3.9.0
python-dotenv==0.10.3
six==1.12.0
urllib3==1.25.3
webassets==0.12.1
Werkzeug==0.15.6
WTForms==2.2.1
